Refactor: Key Toggle Manager
#2

~ 단축키로 컴포넌트 컨트롤 기능 리팩토링
