+++ TotheNorth/Assets/Prefabs/Modal/Contents/Grid/GridWindowPrefab.prefab
@@ -313,7 +313,7 @@ GameObject:
   - component: {fileID: 76761272101309954}
   - component: {fileID: 1856372707341643920}
   m_Layer: 5
-  m_Name: ContentGridPrefab
+  m_Name: GridWindowPrefab
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -392,5 +392,5 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   sizeToUnit: {x: 4, y: 6}
+  keyToToggle: i
   slotPrefab: {fileID: 2028720023028785112, guid: 6daff730708aad941a274c818d4f95e6, type: 3}
-  contentPrefab: {fileID: 0}
