+++ TotheNorth/Assets/Prefabs/Modal/Contents/TestWindowPrefab.prefab
@@ -696,7 +696,7 @@ GameObject:
   - component: {fileID: 76761272101309953}
   - component: {fileID: 76761272101309954}
   m_Layer: 5
-  m_Name: ContentTestPrefab
+  m_Name: TestWindowPrefab
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -737,6 +737,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   sizeToUnit: {x: 4, y: 7}
+  keyToToggle: a
 --- !u!222 &76761272101309953
 CanvasRenderer:
   m_ObjectHideFlags: 0
