+++ TotheNorth/Assets/Scripts/Commons/Abstracts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 6daff730708aad941a274c818d4f95e6
-PrefabImporter:
+guid: f5b9429e60e5a0d41b91b3d3c0d01d45
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
