+++ TotheNorth/Assets/Scripts/Commons/Abstracts/MonoBehaviourControllByKey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c1d5da296a7d124a90eb61edd62c6b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
