+++ TotheNorth/Assets/Scripts/Commons/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 385fdaaaa69a3eb408ebf19ad99300b7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
