+++ TotheNorth/Assets/Scripts/Commons/Interfaces/IControllByKey.cs
@@ -1,0 +1,4 @@
+public interface IControllByKey
+{
+    public void ControllByKey(int purpose);
+}
