+++ TotheNorth/Assets/Scripts/Commons/Interfaces/IControllByKey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6ef5312014c121241bcea84f3430884b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
