+++ TotheNorth/Assets/Scripts/Components/Modals/Interfaces/IContentModalContentController.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IContentModalContentController
 {
     public void ClearContent();
-    public void Toggle(int purpose);
+    public void InitComposition();
 }
